test: cmek tests
