+++ provider/resource_rediscloud_pro_subscription_test.go
@@ -722,30 +722,6 @@ func testAccCheckProSubscriptionDestroy(s *terraform.State) error {
 	return nil
 }
 
-func TestAccResourceRedisCloudProSubscription_CMEK(t *testing.T) {
-
-	//testAccRequiresEnvVar(t, "EXECUTE_TESTS")
-
-	name := acctest.RandomWithPrefix(testResourcePrefix)
-	const resourceName = "rediscloud_subscription.example"
-	testCloudAccountName := os.Getenv("AWS_TEST_CLOUD_ACCOUNT_NAME")
-
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t); testAccAwsPreExistingCloudAccountPreCheck(t) },
-		ProviderFactories: providerFactories,
-		CheckDestroy:      testAccCheckProSubscriptionDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: fmt.Sprintf(testAccResourceRedisCloudProSubscriptionCmekEnabled_part1, testCloudAccountName, name),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName, "name", name),
-					resource.TestCheckResourceAttr(resourceName, "cloud_provider.0.region.0.preferred_availability_zones.#", "1"),
-				),
-			},
-		},
-	})
-}
-
 // TF config for provisioning a new subscription.
 const testAccResourceRedisCloudProSubscription = `
 data "rediscloud_payment_method" "card" {
@@ -790,78 +766,6 @@ resource "rediscloud_subscription" "example" {
     throughput_measurement_by = "operations-per-second"
     throughput_measurement_value = 10000
     modules = ["RedisJSON", "RedisBloom", "RediSearch"]
-  }
-}
-`
-
-const testAccResourceRedisCloudProSubscriptionCmekEnabled_part1 = `
-data "rediscloud_payment_method" "card" {
-  card_type = "Visa"
-}
-
-data "rediscloud_cloud_account" "account" {
-  exclude_internal_account = true
-  provider_type = "AWS" 
-  name = "%s"
-}
-
-resource "rediscloud_subscription" "example" {
-
-  name = "%s"
-  payment_method = "credit-card"
-  payment_method_id = data.rediscloud_payment_method.card.id
-  memory_storage = "ram"
-  cmek_enabled = true
-
-  allowlist {
-    cidrs = ["192.168.0.0/16"]
-    security_group_ids = []
-  }
-
-  cloud_provider {
-    provider = data.rediscloud_cloud_account.account.provider_type
-    cloud_account_id = data.rediscloud_cloud_account.account.id
-    region {
-      region = "eu-west-1"
-      networking_deployment_cidr = "10.0.0.0/24"
-      preferred_availability_zones = ["eu-west-1a"]
-    }
-  }
-}
-`
-
-const testAccResourceRedisCloudProSubscriptionCmekEnabled_update = `
-data "rediscloud_payment_method" "card" {
-  card_type = "Visa"
-}
-
-data "rediscloud_cloud_account" "account" {
-  exclude_internal_account = true
-  provider_type = "AWS" 
-  name = "%s"
-}
-
-resource "rediscloud_subscription" "example" {
-
-  name = "%s"
-  payment_method = "credit-card"
-  payment_method_id = data.rediscloud_payment_method.card.id
-  memory_storage = "ram"
-  cmek_enabled = true
-
-  allowlist {
-    cidrs = ["192.168.0.0/16"]
-    security_group_ids = []
-  }
-
-  cloud_provider {
-    provider = data.rediscloud_cloud_account.account.provider_type
-    cloud_account_id = data.rediscloud_cloud_account.account.id
-    region {
-      region = "eu-west-1"
-      networking_deployment_cidr = "10.0.0.0/24"
-      preferred_availability_zones = ["eu-west-1a"]
-    }
   }
 }
 `
